Subido casos iniciales producto y promo

+++ include/controladores/ControladorProducto.hh
@@ -18,7 +18,7 @@ private:
 
 public:
   static ControladorProducto *getInstancia();
-  void registrarDatosProductos(std::string, int, int, std::string, TipoProducto, DTVendedor*);
+  void registrarDatosProductos(std::string, float, int, std::string, TipoProducto, DTVendedor *);
   void altaNuevoProducto();
   DTProducto *obtenerProductoDisponible(int);
   std::set<DTProducto *> obtenerProductosDisponibles(); // falta hacerla en el cpp // agregado PD chequearlo.
